+++ tests/test_reproject.py
@@ -252,9 +252,9 @@ class ReprojectDatasetTest(unittest.TestCase):
 
         target_ds = reproject_dataset(source_ds, target_gm, interp_methods="triangular")
         self.assertCountEqual(["temperature", "onedim_data"], list(target_ds.data_vars))
-        self.assertAlmostEqual(target_ds.temperature.values[0, 0, 0], 6216.15, places=2)
+        self.assertAlmostEqual(target_ds.temperature.values[0, 0, 0], 6216.14, places=2)
         self.assertAlmostEqual(
-            target_ds.temperature.values[0, -1, -1], 2848.51, places=2
+            target_ds.temperature.values[0, -1, -1], 2848.50, places=2
         )
         self.assertEqual(
             [2, 5, 5],
@@ -267,9 +267,9 @@ class ReprojectDatasetTest(unittest.TestCase):
 
         target_ds = reproject_dataset(source_ds, target_gm, interp_methods=1)
         self.assertCountEqual(["temperature", "onedim_data"], list(target_ds.data_vars))
-        self.assertAlmostEqual(target_ds.temperature.values[0, 0, 0], 6216.15, places=2)
+        self.assertAlmostEqual(target_ds.temperature.values[0, 0, 0], 6216.14, places=2)
         self.assertAlmostEqual(
-            target_ds.temperature.values[0, -1, -1], 2848.51, places=2
+            target_ds.temperature.values[0, -1, -1], 2848.50, places=2
         )
         self.assertEqual(
             [2, 5, 5],